+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added new University of Michigan process and mapping for ingestion
 - New directory for test JSON files that will be ingested
 - New script to delete UMP Manifest links from links table
+- New script to parse download requests from S3 log files for UMP books 
 ## Fixed
 - NYPL records not being added due to SQLAlchemy error
 - Bardo CCE API and Hathi DataFiles URL updated